alternate MCI device ID fixes.
The use of alternate device ID (MCIDevID flag in winmm.ini) now opens a waveaudio device to lock a real mci device ID for use with the cdaudio emulation. This is similar to how Dxwnd handles the emulated device ID and should please all games. The imaginary 0xBEEF / 48879 is still left as the default behaviour just in case.