Handle "open cdaudio" return buffer case with if statement
A game could use mciSendString with the following code:
mciSendStringA("open cdaudio", lpszReturnString, sizeof(lpszReturnString), NULL);
In which case the return buffer should be filled with the device id in emulation. This would be necessary if the program expected the return buffer to contain something meaningful. Now added an if statement to avoid returning a string if none was expected. More testing and examples needed before un-commenting the code.